move default_config into server to fix import issue
